HNS_Dashboard_New/backend/requirements.txt
fastapi
uvicorn
pyodbc
pandas
pydantic
python-dotenv
python-multipart
jinja2
sqlalchemy
slowapi
passlib[bcrypt]
python-jose[cryptography]
